refactor(core): split mock.ts into __mocks__/*.ts
