+++ packages/core/src/__mocks__/user.ts
@@ -1,0 +1,96 @@
+import { User, userInfoSelectFields } from '@logto/schemas';
+import pick from 'lodash.pick';
+
+export const mockUser: User = {
+  id: 'foo',
+  username: 'foo',
+  primaryEmail: 'foo@logto.io',
+  primaryPhone: '111111',
+  roleNames: ['admin'],
+  passwordEncrypted: null,
+  passwordEncryptionMethod: null,
+  passwordEncryptionSalt: null,
+  name: null,
+  avatar: null,
+  identities: {
+    connector1: { userId: 'connector1', details: {} },
+  },
+  customData: {},
+};
+
+export const mockUserResponse = pick(mockUser, ...userInfoSelectFields);
+
+export const mockUserList: User[] = [
+  {
+    id: '1',
+    username: 'foo1',
+    primaryEmail: 'foo1@logto.io',
+    primaryPhone: '111111',
+    roleNames: ['admin'],
+    passwordEncrypted: null,
+    passwordEncryptionMethod: null,
+    passwordEncryptionSalt: null,
+    name: null,
+    avatar: null,
+    identities: {},
+    customData: {},
+  },
+  {
+    id: '2',
+    username: 'foo2',
+    primaryEmail: 'foo2@logto.io',
+    primaryPhone: '111111',
+    roleNames: ['admin'],
+    passwordEncrypted: null,
+    passwordEncryptionMethod: null,
+    passwordEncryptionSalt: null,
+    name: null,
+    avatar: null,
+    identities: {},
+    customData: {},
+  },
+  {
+    id: '3',
+    username: 'foo3',
+    primaryEmail: 'foo3@logto.io',
+    primaryPhone: '111111',
+    roleNames: ['admin'],
+    passwordEncrypted: null,
+    passwordEncryptionMethod: null,
+    passwordEncryptionSalt: null,
+    name: null,
+    avatar: null,
+    identities: {},
+    customData: {},
+  },
+  {
+    id: '4',
+    username: 'bar1',
+    primaryEmail: 'bar1@logto.io',
+    primaryPhone: '111111',
+    roleNames: ['admin'],
+    passwordEncrypted: null,
+    passwordEncryptionMethod: null,
+    passwordEncryptionSalt: null,
+    name: null,
+    avatar: null,
+    identities: {},
+    customData: {},
+  },
+  {
+    id: '5',
+    username: 'bar2',
+    primaryEmail: 'bar2@logto.io',
+    primaryPhone: '111111',
+    roleNames: ['admin'],
+    passwordEncrypted: null,
+    passwordEncryptionMethod: null,
+    passwordEncryptionSalt: null,
+    name: null,
+    avatar: null,
+    identities: {},
+    customData: {},
+  },
+];
+
+export const mockUserListResponse = mockUserList.map((user) => pick(user, ...userInfoSelectFields));
