+++ packages/core/src/middleware/koa-user-info.test.ts
@@ -1,6 +1,6 @@
+import { mockUser, mockUserResponse } from '@/__mocks__';
 import RequestError from '@/errors/RequestError';
 import * as userQueries from '@/queries/user';
-import { mockUser, mockUserResponse } from '@/utils/mock';
 import { createContextWithRouteParameters } from '@/utils/test-utils';
 
 import koaUserInfo from './koa-user-info';
